added event.index view for non AJAX request

+++ app/Http/Controllers/CommentsController.php
@@ -34,6 +34,8 @@ class CommentsController extends Controller
      */
     public function destroy(Comment $comment)
     {
+        $this->authorize('update', $comment);
+
         $comment->delete();
 
         return back();
